Moved search files to separate files & folders
